# Conflicts:
#	.idea/misc.xml

+++ app/src/main/java/domain/model/TravelPreview.kt
@@ -5,15 +5,13 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class TravelPreview(
-    val id: Int,
+    val id: Long,
     val name: String,
     val image: String,
 ) : Parcelable {
     companion object {
-        const val ADD_ID = 0
+        const val ADD_ID = 0L
         fun getAddItem() = TravelPreview(ADD_ID, "", "")
         fun List<TravelPreview>.cloned() = ArrayList(this.map { it.copy() })
     }
-
-
 }
